Created using Colab
This project is about predicting loan eligibility using a dataset from Dream Housing Finance company. The goal is to automate the process of determining whether a customer is eligible for a loan based on information they provide in an online application form. This involves building a machine learning model to classify loan applications as either eligible or not eligible.

The dataset contains various features like Gender, Marital Status, Education, Number of Dependents, Income, Loan Amount, Credit History, and others. These features are used to train the model and predict loan eligibility. The project involves analyzing the data to understand the relationships between these features and the target variable (Loan Status), identifying important features, and building a predictive model.